Reduce heterogeneity by inlining TenantApplications, and some small fixes
