+++ configserver/src/main/java/com/yahoo/vespa/config/server/tenant/TenantRepository.java
@@ -20,6 +20,7 @@ import org.apache.curator.framework.state.ConnectionStateListener;
 import org.apache.zookeeper.KeeperException;
 
 import java.time.Duration;
+import java.time.temporal.ChronoUnit;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
@@ -37,6 +38,7 @@ import java.util.concurrent.Future;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.ScheduledThreadPoolExecutor;
 import java.util.concurrent.TimeUnit;
+import java.util.logging.Level;
 import java.util.logging.Logger;
 import java.util.stream.Collectors;
 
@@ -57,7 +59,7 @@ import java.util.stream.Collectors;
  * @author Vegard Havdal
  * @author Ulf Lilleengen
  */
-public class TenantRepository implements ConnectionStateListener, PathChildrenCacheListener {
+public class TenantRepository {
 
     public static final TenantName HOSTED_VESPA_TENANT = TenantName.from("hosted-vespa");
     private static final TenantName DEFAULT_TENANT = TenantName.defaultName();
@@ -104,7 +106,7 @@ public class TenantRepository implements ConnectionStateListener, PathChildrenCa
         this.curator = globalComponentRegistry.getCurator();
         metricUpdater = globalComponentRegistry.getMetrics().getOrCreateMetricUpdater(Collections.emptyMap());
         this.tenantListeners.add(globalComponentRegistry.getTenantListener());
-        curator.framework().getConnectionStateListenable().addListener(this);
+        curator.framework().getConnectionStateListenable().addListener(this::stateChanged);
 
         curator.create(tenantsPath);
         createSystemTenants(configserverConfig);
@@ -113,7 +115,7 @@ public class TenantRepository implements ConnectionStateListener, PathChildrenCa
         if (useZooKeeperWatchForTenantChanges) {
             this.directoryCache = Optional.of(curator.createDirectoryCache(tenantsPath.getAbsolute(), false, false, pathChildrenExecutor));
             this.directoryCache.get().start();
-            this.directoryCache.get().addListener(this);
+            this.directoryCache.get().addListener(this::childEvent);
         } else {
             this.directoryCache = Optional.empty();
         }
@@ -318,8 +320,7 @@ public class TenantRepository implements ConnectionStateListener, PathChildrenCa
         return ret.toString();
     }
 
-    @Override
-    public void stateChanged(CuratorFramework framework, ConnectionState connectionState) {
+    private void stateChanged(CuratorFramework framework, ConnectionState connectionState) {
         switch (connectionState) {
             case CONNECTED:
                 metricUpdater.incZKConnected();
@@ -339,8 +340,7 @@ public class TenantRepository implements ConnectionStateListener, PathChildrenCa
         }
     }
 
-    @Override
-    public void childEvent(CuratorFramework framework, PathChildrenCacheEvent event) {
+    private void childEvent(CuratorFramework framework, PathChildrenCacheEvent event) {
         switch (event.getType()) {
             case CHILD_ADDED:
             case CHILD_REMOVED:
@@ -351,8 +351,16 @@ public class TenantRepository implements ConnectionStateListener, PathChildrenCa
 
     public void close() {
         directoryCache.ifPresent(Curator.DirectoryCache::close);
-        pathChildrenExecutor.shutdown();
-        checkForRemovedApplicationsService.shutdown();
+        try {
+            pathChildrenExecutor.shutdown();
+            checkForRemovedApplicationsService.shutdown();
+            pathChildrenExecutor.awaitTermination(50, TimeUnit.SECONDS);
+            checkForRemovedApplicationsService.awaitTermination(50, TimeUnit.SECONDS);
+        }
+        catch (InterruptedException e) {
+            log.log(Level.WARNING, "Interrupted while shutting down.", e);
+            Thread.currentThread().interrupt();
+        }
     }
 
     public boolean checkThatTenantExists(TenantName tenant) {
