+++ configserver/src/test/java/com/yahoo/vespa/config/server/session/LocalSessionTest.java
@@ -15,8 +15,9 @@ import com.yahoo.config.provision.TenantName;
 import com.yahoo.path.Path;
 import com.yahoo.slime.Slime;
 import com.yahoo.transaction.NestedTransaction;
+import com.yahoo.vespa.config.server.MockReloadHandler;
 import com.yahoo.vespa.config.server.SuperModelGenerationCounter;
-import com.yahoo.vespa.config.server.application.MemoryTenantApplications;
+import com.yahoo.vespa.config.server.application.TenantApplications;
 import com.yahoo.vespa.config.server.deploy.DeployHandlerLogger;
 import com.yahoo.vespa.config.server.deploy.TenantFileSystemDirs;
 import com.yahoo.vespa.config.server.deploy.ZooKeeperClient;
@@ -200,7 +201,7 @@ public class LocalSessionTest {
                         FilesApplicationPackage.fromFile(testApp),
                         zkc,
                         sessionDir,
-                        new MemoryTenantApplications(),
+                        TenantApplications.create(curator, new MockReloadHandler(), tenant),
                         new HostRegistry<>(),
                         superModelGenerationCounter,
                         flagSource));
