+++ configserver/src/test/java/com/yahoo/vespa/config/server/tenant/TenantTest.java
@@ -3,19 +3,22 @@ package com.yahoo.vespa.config.server.tenant;
 
 import com.google.common.testing.EqualsTester;
 import com.yahoo.config.provision.TenantName;
+import com.yahoo.vespa.config.server.MockReloadHandler;
 import com.yahoo.vespa.config.server.TestComponentRegistry;
-import com.yahoo.vespa.config.server.application.MemoryTenantApplications;
+import com.yahoo.vespa.config.server.application.TenantApplications;
+import com.yahoo.vespa.curator.mock.MockCurator;
 import org.junit.Before;
 import org.junit.Test;
 
 import static org.hamcrest.CoreMatchers.not;
 import static org.hamcrest.Matchers.is;
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertThat;
 
 /**
  * @author Ulf Lilleengen
  */
 public class TenantTest {
+
     private final TestComponentRegistry componentRegistry = new TestComponentRegistry.Builder().build();
 
     private Tenant t1;
@@ -35,7 +38,7 @@ public class TenantTest {
         TenantRepository tenantRepository = new TenantRepository(componentRegistry, false);
         TenantName tenantName = TenantName.from(name);
         TenantBuilder tenantBuilder = TenantBuilder.create(componentRegistry, tenantName)
-                .withApplicationRepo(new MemoryTenantApplications());
+                .withApplicationRepo(TenantApplications.create(new MockCurator(), new MockReloadHandler(), tenantName));
         tenantRepository.addTenant(tenantBuilder);
         return tenantRepository.getTenant(tenantName);
     }
@@ -56,11 +59,4 @@ public class TenantTest {
         assertThat(t1.hashCode(), is(not(t4.hashCode())));
     }
 
-    @Test
-    public void close() {
-        MemoryTenantApplications repo = (MemoryTenantApplications) t1.getApplicationRepo();
-        assertTrue(repo.isOpen());
-        t1.close();
-        assertFalse(repo.isOpen());
-    }
 }
